backend/requirements.txt
Flask
Flask-CORS
PyJWT
python-dotenv
SQLAlchemy
Werkzeug
